Moved table_row and header styles to theme. The implemented renderer for the theme also lives there.
