+++ crates/core/Cargo.toml
@@ -36,7 +36,7 @@ grin_wallet_controller = { git = "https://github.com/mimblewimble/grin-wallet", 
 grin_wallet_api = { git = "https://github.com/mimblewimble/grin-wallet", branch = "master" }
 grin_wallet_impls = { git = "https://github.com/mimblewimble/grin-wallet", branch = "master" }
 grin_wallet_libwallet = { git = "https://github.com/mimblewimble/grin-wallet", branch = "master" }
-
+grin-gui-widgets = {path = "../widgets"}
 
 
 ############ Local testing ################
@@ -86,7 +86,6 @@ backtrace = "0.3"
 lazy_static = "1"
 
 iced = { version = "0.5.2", default-features = false }
-#iced = { path = "../../../iced"}
 iced_futures = { version = "0.5.1", features = ["async-std"] }
 iced_style = "0.5.0"
 iced_graphics = { version = "0.4.0" }
