+++ crates/widgets/Cargo.toml
@@ -20,4 +20,4 @@ iced_graphics = { version = "0.4.0" }
 iced_core = { version = "0.6.1" }
 iced_wgpu = { version = "0.6.1", optional = true }
 iced_glow = { version = "0.4.1", optional = true }
-grin-gui-core = { path = "../core"}
+#grin-gui-core = { path = "../core"}
