--- crates/widgets/src/renderer.rs
@@ -1,2 +1,0 @@
-mod header;
-mod table_row;
