+++ src/gui/element/wallet/operation/create_tx.rs
@@ -5,7 +5,7 @@ use grin_gui_core::{
     config::Config,
     wallet::{TxLogEntry, TxLogEntryType},
 };
-use grin_gui_widgets::{header, Header, TableRow};
+use grin_gui_widgets::{header};
 //use iced::button::StyleSheet;
 use iced_aw::Card;
 use iced_native::Widget;
@@ -22,7 +22,7 @@ use {
     grin_gui_core::wallet::{StatusMessage, WalletInfo, WalletInterface},
     grin_gui_core::{node::amount_to_hr_string, theme::ColorPalette},
     iced::{alignment, Alignment, Command, Length},
-    grin_gui_core::theme::{Container, Button, Element, Column, PickList, Row, Scrollable, Text, TextInput},
+    grin_gui_core::theme::{Container, Button, Element, Column, PickList, Row, Scrollable, Text, TextInput, Header, TableRow},
     iced::widget::{
         button, pick_list, scrollable, text_input, Checkbox, Space,
     },
